+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/core/providers/GameStateProviderTemplate.java
@@ -53,6 +53,7 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
 
     // Combat events
     public static final EventTemplateV1 E_WARDEN_NEARBY = new EventTemplateV1("warden_nearby", instance -> instance::wardenNearby, McVersion.V1_19);
+    public static final EventTemplateV1 E_IS_TARGETED = new EventTemplateV1("is_targeted", instance -> instance::isTargeted);
     public static final EventTemplateV1 E_IN_COMBAT = new EventTemplateV1("in_combat", instance -> instance::inCombat);
     public static final EventTemplateV1 E_BOSS_FIGHT = new EventTemplateV1("boss_fight", instance -> instance::inBossFight);
 
@@ -60,10 +61,10 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
             E_MAIN_MENU, E_JOINING, E_DISCONNECTED, E_CREDITS, E_PAUSED, E_IN_GAME,
             E_DAY, E_DAWN, E_DUSK, E_NIGHT,
             E_DOWNFALL, E_RAIN, E_SNOW, E_THUNDERING,
-            E_VILLAGE, E_RANCH, E_WARDEN_NEARBY,
+            E_VILLAGE, E_RANCH,
             E_DEAD, E_SLEEPING, E_FISHING, E_UNDER_WATER, E_IN_LAVA, E_ON_FIRE, E_IN_POWDER_SNOW, E_DROWNING,
             E_MINECART, E_BOAT, E_HORSE, E_DONKEY, E_PIG, E_ELYTRA,
-            E_IN_COMBAT, E_BOSS_FIGHT
+            E_WARDEN_NEARBY, E_IS_TARGETED, E_IN_COMBAT, E_BOSS_FIGHT
     };
 
 
@@ -149,7 +150,6 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
     // Location events
     public abstract BoolVal inVillage();
     public abstract BoolVal inRanch();
-    public abstract BoolVal wardenNearby();
 
 
     // ------------------------------------------------------------------------------------------------
@@ -176,6 +176,8 @@ public abstract class GameStateProviderTemplate extends BaseGameStateProvider
 
     // ------------------------------------------------------------------------------------------------
     // Combat events
+    public abstract BoolVal wardenNearby();
+    public abstract BoolVal isTargeted();
     public abstract BoolVal inCombat();
     public abstract BoolVal inBossFight();