+++ v1_18_2/src/main/java/me/molybdenum/ambience_mini/v1_18_2/client/core/util/Notification.java
@@ -28,15 +28,10 @@ public class Notification extends BaseNotification<Component>
     }
 
     @Override
-    protected void addOrUpdateToast(Component message) {
-        SystemToast systemtoast = mc.getToasts().getToast(SystemToast.class, TUTORIAL_HINT);
-        if (systemtoast == null) {
-            mc.getToasts().addToast(SystemToast.multiline(
-                    Minecraft.getInstance(), TUTORIAL_HINT, title, message
-            ));
-        } else {
-            systemtoast.reset(title, message);
-        }
+    protected void addToast(Component message) {
+        mc.getToasts().addToast(SystemToast.multiline(
+                Minecraft.getInstance(), TUTORIAL_HINT, title, message
+        ));
     }
 
     @Override