+++ v1_18_2/src/main/java/me/molybdenum/ambience_mini/v1_18_2/mixin/BossHealthOverlayMixin.java
@@ -1,0 +1,25 @@
+package me.molybdenum.ambience_mini.v1_18_2.mixin;
+
+import com.google.common.collect.Maps;
+import me.molybdenum.ambience_mini.engine.client.core.state.VolumeState;
+import net.minecraft.client.gui.components.BossHealthOverlay;
+import net.minecraft.client.gui.components.LerpingBossEvent;
+import net.minecraft.client.sounds.SoundManager;
+import net.minecraft.sounds.SoundSource;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+
+@Mixin(BossHealthOverlay.class)
+public class BossHealthOverlayMixin {
+    @Final
+    @Shadow
+    final Map<UUID, LerpingBossEvent> events = Maps.newConcurrentMap();
+}