+++ v1_21_1/src/main/java/me/molybdenum/ambience_mini/v1_21_1/client/core/state/CombatState.java
@@ -16,6 +16,7 @@ import net.minecraft.world.phys.Vec3;
 import net.neoforged.fml.util.ObfuscationReflectionHelper;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -32,7 +33,7 @@ public class CombatState extends BaseCombatState<Entity, Vec3>
 
     @Override
     public int getEntityId(Entity entity) {
-        return 0;
+        return entity.getId();
     }
 
     @Override
@@ -66,10 +67,8 @@ public class CombatState extends BaseCombatState<Entity, Vec3>
         Map<UUID, LerpingBossEvent> bossMap =
                 ObfuscationReflectionHelper.getPrivateValue(BossHealthOverlay.class, mc.gui.getBossOverlay(), "events");
 
-        return bossMap == null
-                ? List.of()
-                : bossMap.values()
-                .stream()
+        List<LerpingBossEvent> bosses = bossMap == null ? List.of() : new ArrayList<>(bossMap.values());
+        return bosses.stream()
                 .map(bossEvent -> extractBossKeyOrName(bossEvent.getName().getContents()))
                 .toList();
     }